requirements.txt
python-magic~=0.4.15
google-api-python-client~=1.7.7
google-cloud-storage~=1.13.2
google-auth-httplib2~=0.0.3
google-auth~=1.6.2